refactor(FFmpegConverterTest): 예외상황 테스트 추가

+++ backend/src/test/java/com/wooteco/nolto/image/infrastructure/FFmpegConverterTest.java
@@ -1,5 +1,6 @@
 package com.wooteco.nolto.image.infrastructure;
 
+import com.wooteco.nolto.exception.InternalServerErrorException;
 import com.wooteco.nolto.image.config.FFmpegConfig;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
@@ -11,8 +12,10 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.util.Objects;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SpringBootTest(classes = {FFmpegConverter.class, FFmpegConfig.class})
 class FFmpegConverterTest {
@@ -30,7 +33,9 @@ class FFmpegConverterTest {
 
     private void 변환_후_생성된_파일삭제() throws IOException {
         URL resource = getClass().getClassLoader().getResource("static/" + mp4FileName);
-        Files.delete(Paths.get(resource.getPath()));
+        if (Objects.nonNull(resource)) {
+            Files.delete(Paths.get(resource.getPath()));
+        }
     }
 
     @DisplayName("ffmpeg 명령어를 사용해서 gif파일을 mp4 파일로 변환한다.")
@@ -46,7 +51,36 @@ class FFmpegConverterTest {
         // when
         ffmpegConverter.convertGifToMp4(gifFilePath, mp4FilePath);
 
+        // then
         URL mp4URL = getClass().getClassLoader().getResource("static/" + mp4FileName);
         assertThat(mp4URL.getPath()).isEqualTo(mp4FilePath);
+    }
+
+    @DisplayName("gif파일이 파일 경로에 존재하지 않으면 예외가 발생한다. ")
+    @Test
+    void convertGifToMp4WithInvalidGifFilePath() {
+        // given
+        String gifFilePath = "static/매우잘못된경로에있는파일.gif";
+        int indexOfExtensionDot = gifFilePath.lastIndexOf(".");
+        String filePathWithoutExtension = gifFilePath.substring(0, indexOfExtensionDot);
+        String mp4FilePath = filePathWithoutExtension + ".mp4";
+
+        // when then
+        assertThatThrownBy(() -> ffmpegConverter.convertGifToMp4(gifFilePath, mp4FilePath))
+                .isInstanceOf(InternalServerErrorException.class)
+                .hasMessage("gif파일을 mp4파일로 변환에 실패하였습니다.");
+    }
+
+    @DisplayName("gif파일이 파일 경로에 존재하지 않으면 예외가 발생한다. ")
+    @Test
+    void convertGifToMp4WithInvalidMp4FilePath() {
+        // given
+        URL resource = getClass().getClassLoader().getResource("static/" + gifFileName);
+        String gifFilePath = resource.getPath();
+        String mp4FilePath = "static/이런경로가없는데/매우잘못된경로에있는파일.mp4";
+
+        // when then
+        assertThatThrownBy(() -> ffmpegConverter.convertGifToMp4(gifFilePath, mp4FilePath))
+                .isInstanceOf(RuntimeException.class);
     }
 }
